requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
motor==3.7.1
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
